+++ EssentialFeed/EssentialFeed/Feed Cache/Infrastructure/CoreData/CoreDataHelpers.swift
@@ -1,0 +1,42 @@
+//
+//  CoreDataHelpers.swift
+//  EssentialFeed
+//
+//  Created by panditpakhurde on 28/05/20.
+//  Copyright © 2020 Quikr. All rights reserved.
+//
+
+import CoreData
+
+internal extension NSPersistentContainer {
+    
+    enum LoadingError: Swift.Error {
+        case modelNotFound
+        case failedToLoadPersistenStores(Swift.Error)
+    }
+    
+    static func load(modelName name: String, url: URL, in bundle: Bundle) throws -> NSPersistentContainer {
+        guard let model = NSManagedObjectModel.with(name: name, in: bundle) else {
+            throw LoadingError.modelNotFound
+        }
+        
+        let description = NSPersistentStoreDescription(url: url)
+        let container = NSPersistentContainer(name: name, managedObjectModel: model)
+        container.persistentStoreDescriptions = [description]
+        
+        var loadError: Swift.Error?
+        container.loadPersistentStores {    loadError = $1    }
+        
+        try loadError.map { throw LoadingError.failedToLoadPersistenStores($0) }
+        
+        return container
+    }
+}
+
+private extension NSManagedObjectModel {
+    static func with(name: String, in bundle: Bundle) -> NSManagedObjectModel? {
+        return bundle
+            .url(forResource: name, withExtension: "momd")
+            .flatMap { (NSManagedObjectModel(contentsOf: $0)) }
+    }
+}
